Ported read-only test
