+++ enginetest/enginetests.go
@@ -17,6 +17,7 @@ package enginetest
 import (
 	"context"
 	"github.com/liquidata-inc/go-mysql-server"
+	"github.com/liquidata-inc/go-mysql-server/auth"
 	"github.com/liquidata-inc/go-mysql-server/sql"
 	"github.com/liquidata-inc/go-mysql-server/sql/analyzer"
 	"github.com/liquidata-inc/go-mysql-server/sql/parse"
@@ -148,6 +149,43 @@ func TestOrderByGroupBy(t *testing.T, harness Harness) {
 		"SELECT team, COUNT(*) FROM members GROUP BY team ORDER BY columndoesnotexist",
 	)
 	require.Error(err)
+}
+
+func TestReadOnly(t *testing.T, harness Harness) {
+	require := require.New(t)
+
+	db := harness.NewDatabase("mydb")
+	harness.NewTable(db, "mytable", sql.Schema{
+		{Name: "i", Type: sql.Int64, Source: "mytable"},
+		{Name: "s", Type: sql.Text, Source: "mytable"},
+	})
+
+	catalog := sql.NewCatalog()
+	catalog.AddDatabase(db)
+
+	au := auth.NewNativeSingle("user", "pass", auth.ReadPerm)
+	cfg := &sqle.Config{Auth: au}
+	a := analyzer.NewBuilder(catalog).Build()
+	e := sqle.New(catalog, a, cfg)
+	idxReg := sql.NewIndexRegistry()
+
+	_, _, err := e.Query(NewCtx(idxReg), `SELECT i FROM mytable`)
+	require.NoError(err)
+
+	writingQueries := []string{
+		`CREATE INDEX foo USING BTREE ON mytable (i, s)`,
+		`CREATE INDEX foo USING pilosa ON mytable (i, s)`,
+		`DROP INDEX foo ON mytable`,
+		`INSERT INTO mytable (i, s) VALUES(42, 'yolo')`,
+		`CREATE VIEW myview AS SELECT i FROM mytable`,
+		`DROP VIEW myview`,
+	}
+
+	for _, query := range writingQueries {
+		_, _, err = e.Query(NewCtx(idxReg), query)
+		require.Error(err)
+		require.True(auth.ErrNotAuthorized.Is(err))
+	}
 }
 
 var pid uint64
